moved all python files allociated with the database creation into a separate folder; cleaned out unused code
